+++ Makefile
@@ -1,6 +1,6 @@
 default: | help
 
-SRC_DIR=ptmpdy
+SRC_DIR=sls_enum
 
 format: ## format codes
 	poetry run autoflake -ri --remove-all-unused-imports --ignore-init-module-imports --remove-unused-variables ${SRC_DIR} && \
