+++ sls_enum/__init__.py
